requirements.txt
deepgram-sdk==3.2.7
python-dotenv==1.0.1
langchain==0.1.9
langchain_groq==0.1.4
numpy==1.26.4
openai==1.30.3
PyAudio==0.2.14
pydantic==2.6.2
requests==2.32.2
langchain_openai==0.1.7
elevenlabs==1.2.2
python-dotenv==1.0.1
crewai==0.30.11
duckduckgo-search==6.1.7